backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.0
python-multipart==0.0.12

# ── Validation & settings ─────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ── Database ──────────────────────────────────────────────────────────────
sqlalchemy==2.0.35
aiosqlite==0.20.0        # SQLite async driver (default/dev)
asyncpg==0.29.0          # PostgreSQL async driver
aiomysql==0.2.0          # MySQL async driver

# ── Cache / session ───────────────────────────────────────────────────────
redis==5.1.1

# ── AI / Agents ───────────────────────────────────────────────────────────
# NOTE: anthropic removed — migrated to Groq (llama-3.3-70b-versatile)
groq==0.11.0
openai==1.47.0           # Whisper STT (via Groq API compatible endpoint)
langgraph==0.2.28        # Multi-agent orchestration
chromadb==0.5.15         # Vector DB for persistent session memory
sentence-transformers==3.1.1 # Local embeddings for vector memory

# ── Data processing ───────────────────────────────────────────────────────
duckdb==1.1.0
pandas==2.2.3
openpyxl==3.1.5          # Excel read/write
numpy==1.26.4

# ── File I/O ──────────────────────────────────────────────────────────────
aiofiles==24.1.0

# ── HTTP client (for external APIs) ──────────────────────────────────────
httpx==0.27.2

# ── Auth ─────────────────────────────────────────────────────────────────
PyJWT==2.9.0             # JWT token generation and validation
passlib[bcrypt]==1.7.4   # Password hashing

# ── Export ───────────────────────────────────────────────────────────────
reportlab==4.2.5         # PDF report generation
